Move `HsBindgen.SHs` to `HsBindgen.Backend.SHs`
