+++ hs-bindgen/src-internal/HsBindgen/Backend/Artefact/TH/Translation.hs
@@ -39,7 +39,7 @@ import HsBindgen.Imports
 import HsBindgen.Language.C qualified as C
 import HsBindgen.Language.Haskell
 import HsBindgen.NameHint
-import HsBindgen.SHs.AST
+import HsBindgen.Backend.SHs.AST
 
 import HsBindgen.Runtime.IncompleteArray     qualified
 import HsBindgen.Runtime.Bitfield            qualified
